improving timeouts
changes:
- add SolveState::TIMEOUT (3) to the Python interface.
- add timeout parameter to propagated and pruneDomains from Python
interface.
- SolveState::TIMEOUT is now returned by runFull when timeout is
detected to be exceeded. In particularly hard problems, timeouts may
take a while to be detected (as we only check the outer loop). The
search state is still valid after timeout is reached.
- propagate and pruneDomains no longer throw an unsatEncounter exception
when called on an unsatisfiable problem, but have a defined return
value.

rationale:
It is useful to have some form of recoverable timeout from the Python
interface, even though it is not as finegrained as the non-recoverable
timeout from the commandline.